Added region-of-interest support to replace center point
- The user can now specify a bounding rectangle (sw and ne corner
points) instead of a center point, which I think is more intuitive and
probably what we want.

- The initial view is now the tightest possible bound that Maps will
give that includes the full rectangle.  i.e., it is as zoomed in as
possible.

- The panning restriction is updated to work with the region of
interest.  So, when the user is panning, the center of map can never go
outside the ROI box we specify.  (Previously, the user was restricted
to an arbitrary bounds.)  One option is never to let the user pan out
of that ROI box at all if they're zoomed in enough, but then the
panning model kind of breaks/changes at that initial zoom level
(because the initial map view is not guaranteed to be exactly the ROI)
or if the user wants to zoom out more for whatever reason.

- Overlays for ROI and debugging